Extract NodeError from node/src/lib.rs
No semantic changes.
